+++ functions/token.js
@@ -21,11 +21,43 @@ const headers = {
   'Content-Type': 'application/json'
 };
 
-module.exports.handler = (event, context, callback) => {
+/**
+ * The various secret constants to access the Twilio API are stored in sereral
+ * places. The file /.netlify.toml holds the TWILIO_API_KEYS_ENV values. They
+ * are set based on the build context when Netlify makes a build.
+ * (https://docs.netlify.com/site-deploys/overview/#deploy-contexts)
+ *
+ * The TWILIO_ACCOUNT_SID, TWILIO_API_KEY_SID, and TWILIO_API_KEY_SECRET
+ * are stored inside the Netlify admin panel. They are injected during build.
+ * Also, during local development, they are pulled down from the servers. So,
+ * your local app (netlify dev), has access to the same API keys as when a
+ * build is made on the Netlify servers.
+ */
 
-  const TWILIO_ACCOUNT_SID = process.env.TWILIO_ACCOUNT_SID;
-  const TWILIO_API_KEY_SID = process.env.TWILIO_API_KEY_SID;
-  const TWILIO_API_KEY_SECRET = process.env.TWILIO_API_KEY_SECRET;
+// the Twilio Account SID is universal for all services
+const TWILIO_ACCOUNT_SID = process.env.TWILIO_ACCOUNT_SID;
+// This value comes from netlify.toml, and is a based on Netlify build context
+// It is either "production" or "development"
+const TWILIO_API_KEYS_ENV = process.env.TWILIO_API_KEYS_ENV;
+// These API keys were generated inside the Twilio admin panel, then are stored
+// inside the Netlify admin panel.
+const TWILIO_API_KEYS = {
+  development: {
+    TWILIO_API_KEY_SID:    process.env.TWILIO_API_KEY_SID_DEV,
+    TWILIO_API_KEY_SECRET: process.env.TWILIO_API_KEY_SECRET_DEV,
+  },
+  production: {
+    TWILIO_API_KEY_SID:    process.env.TWILIO_API_KEY_SID_PROD,
+    TWILIO_API_KEY_SECRET: process.env.TWILIO_API_KEY_SECRET_PROD,
+  }
+}
+
+// If you have access to the Netlify environment context, use it to determien
+// which API keys to use. Otherwise, fallback to the 'development' API keys.
+const TWILIO_API_KEY_SID =    TWILIO_API_KEYS[TWILIO_API_KEYS_ENV] ? TWILIO_API_KEYS[TWILIO_API_KEYS_ENV].TWILIO_API_KEY_SID :    TWILIO_API_KEYS['development'].TWILIO_API_KEY_SID;
+const TWILIO_API_KEY_SECRET = TWILIO_API_KEYS[TWILIO_API_KEYS_ENV] ? TWILIO_API_KEYS[TWILIO_API_KEYS_ENV].TWILIO_API_KEY_SECRET : TWILIO_API_KEYS['development'].TWILIO_API_KEY_SECRET;
+
+module.exports.handler = (event, context, callback) => {
 
   // We only care about POSTs with body data
   if(event.httpMethod !== 'POST' || !event.body) {